Separate config into its own package
