+++ gosuper/gosuper/main.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"resin-supervisor/gosuper/application"
+	"resin-supervisor/gosuper/config"
 	"resin-supervisor/gosuper/device"
 	"resin-supervisor/gosuper/psutils"
 	"resin-supervisor/gosuper/supermodels"
@@ -49,15 +50,15 @@ func main() {
 	hostproc := os.Getenv("HOST_PROC")
 	defer startOOMProtectionTimer(hostproc, dockerSocket).Stop()
 
-	config := GetSupervisorConfig()
+	superConfig := config.GetSupervisorConfig()
 	go connectivityCheck()
-	if appsCollection, dbConfig, err := supermodels.New(config.DatabasePath); err != nil {
+	if appsCollection, dbConfig, err := supermodels.New(superConfig.DatabasePath); err != nil {
 		log.Fatal("Failed to start database")
 	} else if theDevice, err := device.New(appsCollection, dbConfig); err != nil {
 		log.Fatal("Failed to start device bootstrapping")
 	} else if applicationManager, err := application.NewManager(appsCollection, dbConfig, theDevice); err != nil {
 		log.Fatal("Failed to initialize applications manager")
-	} else if err = StartApi(config.ListenPort, applicationManager); err != nil {
+	} else if err = StartApi(superConfig.ListenPort, applicationManager); err != nil {
 		log.Fatal("Failed to initialize Supervisor API")
 	}
 }
