+++ gosuper/test_gosuper.sh
@@ -2,3 +2,4 @@
 set -e
 go test -v ./gosuper
 go test -v ./supermodels
+go test -v ./config
